colexec: extract hashTable into a separate file
This commit moves hashTable out of hashjoiner file since the former is
used by other operators as well now. Also, make minor stylistic changes.

Release note: None
