+++ pkg/sql/colexec/.gitignore
@@ -7,6 +7,7 @@ const.eg.go
 count_agg.eg.go
 distinct.eg.go
 hashjoiner.eg.go
+hashtable.eg.go
 like_ops.eg.go
 mergejoinbase.eg.go
 mergejoiner_fullouter.eg.go
